// client/src/app/model/Item.ts
export interface Item {
    name: string;
    type: string;
    categoryId: string;
    tare: string;
    quantity: number;
    weight: number;
    unit: string;
    image: string;
    menuItem: string;
    averageCost: number;
    sellingPrice: number;
    creatorId: string;
    id: string;
    executor: string;
    autoSellingPrice: boolean;
    description: string;
}